#hotfix исправил баг обновления изображения

+++ src/main/java/ru/skypro/homework/service/ImageService.java
@@ -34,7 +34,7 @@ public class ImageService {
 
         try {
             image.setPath(path)
-                    .setSize(file.getSize())
+                    .setSize((int) file.getSize())
                     .setMediaType(file.getContentType())
                     .setData(file.getBytes());
         } catch (IOException e) {
